DGPs, posterior, and posterior predictive
Function oriented programming. Open source.R and specify the path . Run the remaining lines in the script, and the functions found in dgen.R and blr.R will be loaded (the functions loaded are also printed to the console). Use the temp_testing.R script to try out functionality. Sketched builder.R function for the product of two multivariate normal pdfs.